eshop/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.115
botocore==1.29.115
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2
django-advanced-filters==2.0.0
django-braces==1.15.0
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==2.4.0
django-six==1.0.5
django-storages==1.13.2
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
jmespath==1.0.1
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
s3transfer==0.6.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
urllib3==1.26.15
